python_server/requirements.txt
Flask
Pillow
selenium
spotipy